New type system

This MR implements the new type system which is used to choose the best type for the generated integer expressions (in particular, either mpz or a standard C integral type).

The core of the implementation follows Figures 3 & 4 of the following paper : http://julien.signoles.free.fr/publis/2015_jfla.pdf. Sorry it is in French, but the Figures are in an international mathematical language :smile: . The main modification is the integration of a context in order to replace the 2 coercion rules (in order to transform a formal system in which more than 1 rule may be applied into an algorithm).

It is probably not a good idea to read this MR commit by commit since interval.ml* are new files and typing.ml* were fully rewritten. It is probably much better to read these files from scratch and look at the differences between the last master commit and the last commit in this branch for the rest.



See merge request !50

+++ src/plugins/e-acsl/tests/bts/oracle/gen_bts1326.c
@@ -124,7 +124,7 @@ void __gen_e_acsl_atp_NORMAL_computeAverageAccel(ArrayInt *Accel,
     __e_acsl_assert(__gen_e_acsl_valid_read_6,(char *)"RTE",
                     (char *)"atp_NORMAL_computeAverageAccel",
                     (char *)"mem_access: \\valid_read(__gen_e_acsl_at)",8);
-    __e_acsl_assert(*__gen_e_acsl_at == (((((long)(*__gen_e_acsl_at_2)[4] + (long)(*__gen_e_acsl_at_3)[3]) + (long)(*__gen_e_acsl_at_4)[2]) + (long)(*__gen_e_acsl_at_5)[1]) + (long)(*__gen_e_acsl_at_6)[0]) / (long)5,
+    __e_acsl_assert(*__gen_e_acsl_at == (int)((((((*__gen_e_acsl_at_2)[4] + (*__gen_e_acsl_at_3)[3]) + (*__gen_e_acsl_at_4)[2]) + (*__gen_e_acsl_at_5)[1]) + (*__gen_e_acsl_at_6)[0]) / 5),
                     (char *)"Postcondition",
                     (char *)"atp_NORMAL_computeAverageAccel",
                     (char *)"*\\old(AverageAccel) ==\n(((((*\\old(Accel))[4]+(*\\old(Accel))[3])+(*\\old(Accel))[2])+(*\\old(Accel))[1])+(*\n  \\old(Accel))[0])/5",
